backend/requirements.txt
flask
flask-jwt-extended
mysql-connector-python
bcrypt
